[RUI-241]: Move getDimensionFieldName to data util
