+++ src/components/charts/bars/bars.utils.ts
@@ -3,7 +3,8 @@ import { Theme } from '../../../theme/theme.types';
 import { remarkableTheme } from '../../../theme/theme.constants';
 import { ChartData, ChartOptions } from 'chart.js';
 import { getThemeFormatter } from '../../../theme/formatter/formatter.utils';
-import { getDatalabelPercentage, getDimensionFieldName, groupTailAsOther } from '../charts.utils';
+import { getDatalabelPercentage, groupTailAsOther } from '../charts.utils';
+import { getDimensionFieldName } from '../../../utils/data.utils';
 import { getDimensionMeasureColor } from '../../../theme/styles/styles.utils';
 import { getChartColors } from '@embeddable.com/remarkable-ui';
 import { Context } from 'chartjs-plugin-datalabels';
